moved extractor and new interfaces to own class
