+++ src/main/java/edu/ie3/datamodel/models/input/connector/ConnectorInput.java
@@ -5,7 +5,7 @@
 */
 package edu.ie3.datamodel.models.input.connector;
 
-import edu.ie3.datamodel.models.Nodes;
+import edu.ie3.datamodel.io.extractor.Nodes;
 import edu.ie3.datamodel.models.OperationTime;
 import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.NodeInput;
